+++ src/handler/helpers/mod.rs
@@ -1,5 +1,7 @@
 mod api_error;
 mod api_response;
+mod field_error;
 
 pub use api_error::*;
 pub use api_response::*;
+pub use field_error::*;
